+++ myhome-thermo-central.js
@@ -166,6 +166,16 @@ module.exports = function (RED) {
           if (!Array.isArray(frame)) {
             payload.command_received = frame;
           }
+          // MSG1 : add major node configuration info on both returned message
+          msg.mh_nodeConfigInfo = {
+            'name' : config.name ,
+            'topic' : config.topic ,
+            'gateway' : {
+              'name' : gateway.name ,
+              'host' : gateway.host ,
+              'port' : gateway.port
+            }
+          };
           // MSG1 : Add all current node stored values to payload
           payload.state = payloadInfo.state;
           payload.remoteControl = payloadInfo.remoteControl;